fix: wrong type in set_incidence_counter

+++ libmocasinns/include/wang_landau.hpp
@@ -57,7 +57,7 @@ public:
   //! Get-Accessor for the incidence counter
   const HistoType<EnergyType, long long int>& get_incidence_counter() const { return incidence_counter; }
   //! Set-Accessor for the incidence counter
-  void set_incidence_counter(const HistoType<EnergyType, double>& value) { incidence_counter = value; }
+  void set_incidence_counter(const HistoType<EnergyType, long long int>& value) { incidence_counter = value; }
   //! Get-Accessor for the sweep counter
   long get_sweep_counter() const { return sweep_counter; }
